API : Liste des invités

+++ API/API_Footies/Data/DAO/InviteDAO.cs
@@ -64,8 +64,12 @@ namespace API_Footies.Data.DAO
             return modifie;
         }
 
-        public void SupprimerInvite(long id)
+
+
+       public List<Invite> ListInvite()
         {
+            List<Invite> listeInvite = new List<Invite>();
+
             using (SQLiteConnector connection = new SQLiteConnector())
             {
                 if (connection == null)
@@ -74,14 +78,42 @@ namespace API_Footies.Data.DAO
                 }
                 else
                 {
-                    var parameters = new Dictionary<string, object>()
+                    var dataTable = connection.ExecuteQuery("SELECT * FROM Invite");
+                    foreach (DataRow? row in dataTable.Rows)
                     {
-                        {"@Id", id }
-                    };
-                    connection.ExecuteQuery("DELETE FROM Invite WHERE idInvite=@Id", parameters);
+
+                        Invite invite = new Invite((long)row["idInvite"], row["nom"].ToString(), row["prenom"].ToString(), row["NumTel"].ToString(), row["mail"].ToString());
+
+                        listeInvite.Add(invite);
+                    }
                 }
             }
-        }
 
+            return listeInvite;
+        }
+        
+        
+        
+        public void SupprimerInvite(long id)
+        {
+          using (SQLiteConnector connection = new SQLiteConnector())
+          {
+            if (connection == null)
+            {
+              throw new Exception("Erreur de connexion à la base de données");
+            }
+            else
+            {
+               var parameters = new Dictionary<string, object>()
+            {
+                {"@Id", id }
+            };
+            connection.ExecuteQuery("DELETE FROM Invite WHERE idInvite=@Id", parameters);
+        }
     }
+}
+        
+
+
+
 }
